Fix weird crash in e-hal
The symptom was quite odd; the address in e_platform.emem seemed to be
offset by two words (so that emem[0].phy_base returned the value that
should have been in emem[0].ephy_base) and when attempting to load the
slave program, the stack was thrashed and a segfault occurred.

For whatever reason, doing erts_init_slave_io earlier during init fixed
the problem.