Have code_server manage slave module loading
The set of modules loaded for the slave is a subset of the set of
modules loaded for the master. Once a module is loaded for the slave, it
is always updated in both emulators, although there might be a period
during which the master module has been updated but not the slave
module.

The error_handler module now detects if it is called on the slave, and
calls code:ensure_loaded_slave instead, which adds the requested module
to the set of modules that are kept loaded on both.

Any module in the set of modules that are loaded for the master but not
the slave will have its object code cached in code_server. This ensures
that both emulators run the same code. It is not possible to ensure this
by reloading modules in the master as they are loaded in the slave
without reloading sticky modules too. Unfortunately, the init module had
to be changed to cache the object code of any modules it loads before
code_server is running to facilitate this.