Build range tables for the slave emulator
Parameterisation of loader is now good enough that slave_finish_loading
is not required anymore, and neither is slave:finish_loading/1.

Thus, slave and master modules can now be committed at the same time,
which allows code_server to do race-free module upgrades. It does not do
so yet, however.