Allow the killing of slave processes
There are two ways they can be killed. Either they kill themselves
during a proxied BIF, in which case a new field in master_syscall_bif
carries status flags to assert. If this field contains EXITING,
slave_bif will do the preparation that couldn't be done from the master.

If another process kills them, a new slave command, SLAVE_COMMAND_EXIT,
is sent to the slave, which asserts the PENDING_EXIT flag on the slave
side, at which point the exit proceeds as usual.

In addition, the system lockup when trying to shut down a node with
slave processes running has been worked around pending a fix for the
underlying issue.