Fix incorrect hole check errors in slave
They were caused by bad values of Process->last_htop, which was supposed
to be reset on gc, but wasn't synchronised by slave_state. Since it's
not used for anything else, rather than copying that field with
slave_state, we simply reset it slave-side after a GC syscall. It is
still safe to do a hole check on either slave and master since the field
is reset master-side by the GC anyway.