Add erts_refc_decfree
It frees the object if the reference count reaches zero. It can be (and
is) implemented without blocking slave-side.

Since it was the missing puzzle piece to implement the binary matching
instructions, they are included too (however they might crash if given
heap binaries, which are allocated outside shared dram).