Add loading of native code into slave runtime
hipe_unified_loader and several bifs in hipe_bifs have been
parameterised on "Mode", an atom that is either 'master' or 'slave'.

Furthermore, hipe_mkliterals has been modified so that runtime
parameters can be fetched for either the master or slave runtime.
Runtime-specific parameters in hipe_literals.hrl now look like this:

    -define(P_HP, hipe_bifs:get_rts_param(22, get(hipe_target_rts))).

In order to provide this, hipe_mkliterals from the slave is used to
generate a header 'hipe_slave_literals.h' which is included by
'hipe_bif0.c' on the master.