Make HiPE code also call erts_normalise_proc_bin
This should allow HiPE code to match binaries stored outside of shared
memory (by leveraging the on-demand migration by
erts_normalise_proc_bin).

If we were really crazy for performance, we might consider inlining the
test for whether we need normalisation, but for now we are more
concerned with making it correct first. Also, since it is going to cause
spills along the fast path either way, it shouldn't make that much of a
difference.