[HiPE] Workaround for Epiphany-III bug
The bug causes reads to recently written, off-core, memory locations to
occasionally return stale data. Since we can't always move the heap to
local memory, we construct a kind of "fence", and call it whenever we
build something on the heap.

The fence works on the assumptions that writes that take the same route
through the mesh network can't overtake.

The fence /really/ needs to be inlined everywhere (for performance
reasons), but for now it's a primop.