Implement a simple code in local memory
hipe_unified_loader now calls hipe_bifs:cache_insert/2 for each
function, which creates a trampoline and dispatch table, and returns its
entrypoint.

Calls to this entrypoint go via the function hipe_cold_call_c, that
updates the table with how many times that function has been
called. Once it has been called enough, it is copied to local memory,
and the dispatch table is updated to bypass hipe_cold_call.

Currently, there is no eviction, and although there is a way to get
stack maps for cached functions (binary search on the cache table, and
then on a stack descriptor table), it is not yet implemented.