[slave_alloc] Allocate long-lived right-to-left
For now, this change requires the sorted freelist (so long-lived and
standard can share a single one).

This should help the problem of long-lived allocations getting sprinkled
throughout the address space, eventually causing fatal fragmentation of
shared memory.

With a reordering of suites, fully concurrent testing (mods-conc) now
work! The reorder is required because bs_SUITE leaves a slave binary
alive (presumed to be held alive from the HiPE code it loads rather than
by a reference counting bug) slap-bang in the middle of the shared
memory address space. Clearly, the fact that binaries can't "migrate
back" is a problem for memory use and fragmentation.