Improve behaviour of erl_types:t_map/1
It now behaves as the syntactic map construction; later keys can
override earlier ones if they match. This fixes a few uses of this
function from dialyzer which assumed this was the case.

It also discards rather than appends non-singleton keys, since they are
useless and only serve to harm performance anyway.

Also, introduce erl_types:t_map_put/2, which adds an association to a
map type, as if maps:put/3 was called, or M#{X=>Y} was constructed. It
intentionally takes the association as a pair so that it is usable as-is
in a left fold. For example, the expression above can be expressed as
lists:foldl(fun erl_types:t_map_put/2, MT, [{XT,YT}]). Naturally,
maps:put/3 and maps:update/3 was added to erl_bif_types as well.