Loop in local memory while waiting
Once again, attempts to use the IDLE instructions led only to
instability issues. Instead we just execute a loop from local memory to
at least offload memory.

The while loop in schedule(...) caused a major performance regression
for some of the benchmarks together with the software cache, the reason
for which is as of yet unknown. Making it an if with a short "sleep"
seems to address the problem.

Also, "sleeping" during all syscalls (and not just READY) also seemed to
cause performance regressions, even if the sleep was very short.