Enhance map type representation
A map type is now represented as a four-tuple
{Mandatory, Optional, DefaultKey, DefaultValue}. This allows us to
express type constraints on optional arguments, as well as lets HiPE and
Dialyzer infer the absence of keys, improving performance / accuracy.

For a detailed and formal description of what this four-tuple means,
please read the comment in erl_types.

In the interest of avoiding enormous commits, all touches to dialyzer to
make it pass its testcases are not included in this commit.