hipe: Add o0 and o1 to tests
Now that x86 is no longer broken with these optimisation levels, we add
them to the test suite to ensure they do not break again.

Bump timeout to 6min since tests are run twice as many times.

The option set of o1 was changed to all optimisations that run fast on
both big and small programs, incurring only a slight compile time
increase compared to the old set, but with a, presumably, significant
improvement to speed of compiled code.

Change o0 register allocator to linear_scan.