hipe: Fix bug in trap frame allocation wrappers
The trap frame allocation wrappers occasionally call the garbage
collector, even though built-in functions are not supposed to.

On non-{x86,amd64} platforms, HiPE was optimising the BIF wrapper
interface on the basis that BIFs do not GC. So, when
hipe_reserve_beam_trap_frame called the garbage collector, the state in
the PCB was stale and corruption happened.

Now, these particular BIFs are reclassified as GC BIFs.

Unfortunately, in order to do that we needed to introduce a
gc_bif_interface_3 macro in every hipe_$ARCH_bifs.m4 file.