hipe_arm: Be resilient to crappy RTL
The ARM backend crashes if certain RTL optimisations were omitted,
preventing it from being usable at lower optimisation levels.

One of the problems were caused by shift-by-immediate-zero, which wraps
to immediate-32 with some shiftops. TODO: Someplace should be modified
to crash when these are generated so debuging further instances of this
gets easier in the future.