hipe: Reduce memory use of bs_SUITE:bs_construct
Currently, bs_SUITE:bs_construct/1's huge_binaries() is the most memory
hungry test by far in the HiPE suite without good reason, as all the
bugs it aims to test for can be reproduced with any binary over 2^27
bits (limit for fixnums on 32-bit platforms).

We borrow the free_mem/0 helper function from the emulators
bs_construct_SUITE to dynamically adapt the test to the system running
it.