erl_types: Print records consistently with parsing
Instead of omitting record fields that are any() or 'undefined', record
fields are omitted when their actual type is the same as their declared
type.

This makes it consistent with the behaviour of the parser,
removing the risk of user confusion. Additionally, this also makes
Dialyzer's output more terse in many situations, making it easier to
read, especially when larger record types are used.