hipe_range_split: Faster DEF dataflow data struct
The DEF dataflow was using maps as the data structure for sets of temps
as it was faster than ordsets (unlike in liveness dataflow; presumably
due to a larger average set size), however, defset_union/2 was still a
significant bottleneck.

Using the fact that the keys are integers, we construct a hybrid
ordset/bitset data structure that provide the same asymptotic time and
space performance as an ordset, but performs significantly faster than
both ordsets and maps at the union/2 operation in this particular use
case.

In order to provide faster lookup, after dataflow we transform the
orddict into an array, providing logarithmic rather than linear time.