hipe_range_split: Make Edges ADT faster
Edges was keeping a list of all edges and their corresponding livesets,
but was only ever queried for the sum of the weights of all edges from a
certain partition over which a certain temporary was
live (edges_query/3). By changing Edges to just map {partition,temp}
pairs to weights, we can drastically speed it up.