hipe: Inline flatmap construction
Constructions of map values with literal keys are translated to a new
'mkflatmap' primop that accepts a tuple of arity N as its first
argument and N additional arguments.

Right now that translation happens during BEAM to ICode lowering, since
there is no 'put_map_assoc' primop, so doing so during constant
propagation would be less convenient. If a 'put_map_assoc' primop is
introduced, this transform should preferably be moved to the constant
propagator.